+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.2.0 - **Unreleased**
 
 - Improves support for JRuby
+- Adds direct message room mapping to client abstraction
 
 ## 2.1.3 - 2020-09-18